+++ ledger/src/blockstore.rs
@@ -554,10 +554,6 @@ pub fn banking_trace_path(path: &Path) -> PathBuf {
     path.join("banking_trace")
 }
 
-pub fn banking_retrace_path(path: &Path) -> PathBuf {
-    path.join("banking_retrace")
-}
-
 impl Blockstore {
     pub fn ledger_path(&self) -> &PathBuf {
         &self.ledger_path
@@ -565,10 +561,6 @@ impl Blockstore {
 
     pub fn banking_trace_path(&self) -> PathBuf {
         banking_trace_path(&self.ledger_path)
-    }
-
-    pub fn banking_retracer_path(&self) -> PathBuf {
-        banking_retrace_path(&self.ledger_path)
     }
 
     /// Opens a Ledger in directory, provides "infinite" window of shreds